[Refactor/#123] convert package location
