+++ src/main/java/com/example/waggle/global/exception/JwtAuthenticationException.java
@@ -1,4 +1,4 @@
-package com.example.waggle.global.security.exception;
+package com.example.waggle.global.exception;
 
 
 import com.example.waggle.global.payload.code.ErrorStatus;
