+++ src/main/java/com/example/waggle/global/security/filter/JwtExceptionFilter.java
@@ -1,8 +1,8 @@
 package com.example.waggle.global.security.filter;
 
+import com.example.waggle.global.exception.JwtAuthenticationException;
 import com.example.waggle.global.payload.code.ErrorStatus;
 import com.example.waggle.global.security.exception.CustomErrorSend;
-import com.example.waggle.global.security.exception.JwtAuthenticationException;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
