+++ src/main/java/com/example/waggle/global/security/service/TokenServiceImpl.java
@@ -4,9 +4,9 @@ import com.example.waggle.domain.member.entity.Member;
 import com.example.waggle.domain.member.service.MemberQueryService;
 import com.example.waggle.domain.member.service.RedisService;
 import com.example.waggle.global.exception.GeneralException;
+import com.example.waggle.global.exception.JwtAuthenticationException;
 import com.example.waggle.global.exception.handler.AuthenticationHandler;
 import com.example.waggle.global.payload.code.ErrorStatus;
-import com.example.waggle.global.security.exception.JwtAuthenticationException;
 import com.example.waggle.global.security.object.JwtToken;
 import com.example.waggle.web.converter.MemberConverter;
 import com.example.waggle.web.dto.member.MemberRequest;
